Move schemes into separate JSON files
This has the benefits of separating data from code and of having a
language-agnostic specification of schemes (for implementations of
Sanscript in other programming languages).

This also adds a build step integrating the scheme data into the JS file
-- the file that now contains only the code has been moved to a
subdirectory and the build step (npm run dist) outputs to sanscript.js
(so this causes no breaking changes).