Add JSON schema and validation script (#17)
This schema is written to fit the current format of JSON scheme
descriptions. Some further work should be done to improve this format to
be more consistent and to capture some of the details that are currently
expressed in code (e.g. the definition of the Kolkata scheme and
Brahmic/roman).